Move StateSeparator to a dedicated util file
